@@ -1,15 +1,17 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '28.0.3'
+    namespace 'ru.alexsuvorov.paistewiki'
     defaultConfig {
         applicationId "ru.alexsuvorov.paistewiki"
-        minSdkVersion 16
+        minSdkVersion 21
+        compileSdk 34
         //noinspection OldTargetApi
-        targetSdkVersion 27
-        versionCode 21
-        versionName "1.3.9.5"
+        targetSdkVersion 34
+        versionCode 25
+        versionName "1.4.0"
+        vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
         release {
@@ -19,48 +21,35 @@ android {
     }
     bundle {
         language {
-            // Specifies that the app bundle should not support
-            // configuration APKs for language resources. These
-            // resources are instead packaged with each base and
-            // dynamic feature APK.
             enableSplit = false
         }
     }
     sourceSets { main { assets.srcDirs = ['src/main/assets', 'src/main/assets/'] } }
     compileOptions {
-        targetCompatibility JavaVersion.VERSION_1_8
-        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_17
+        sourceCompatibility JavaVersion.VERSION_17
     }
 }
 
-def support = '27.1.1'
-
 dependencies {
-    //Support
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation "com.android.support:appcompat-v7:$support"
-    implementation "com.android.support:design:$support"
-    implementation "com.android.support:support-v4:$support"
-    implementation "com.android.support:support-compat:$support"
-    implementation "com.android.support:cardview-v7:$support"
-    implementation "com.android.support:recyclerview-v7:$support"
-    implementation "com.android.support:customtabs:$support"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.github.smarteist:autoimageslider:1.1.1'
-    implementation 'com.github.bumptech.glide:glide:4.7.1'
 
-    //Database
+    implementation "androidx.activity:activity-ktx:1.6.1"
+    implementation "androidx.appcompat:appcompat:1.6.1"
+    implementation "androidx.annotation:annotation:1.5.0"
+    implementation "androidx.browser:browser:1.5.0"
+    implementation "androidx.core:core-ktx:1.9.0"
+    implementation "androidx.fragment:fragment-ktx:1.5.0-beta01"
+    implementation "androidx.navigation:navigation-ui-ktx:2.5.3"
+
+    implementation 'com.github.bumptech.glide:glide:4.16.0'
+
     implementation 'android.arch.persistence.room:runtime:1.1.1'
     annotationProcessor "android.arch.persistence.room:compiler:1.1.1"
-    //Firebase
-    implementation 'com.google.firebase:firebase-ads:15.0.1'
-    implementation 'com.google.firebase:firebase-core:16.0.3'
-    implementation 'com.google.android.gms:play-services-ads:15.0.1'
-    //Other
-    implementation 'com.google.gms:google-services:4.0.2'
+
+    implementation 'com.google.firebase:firebase-core:21.1.1'
+
+    implementation 'com.google.gms:google-services:4.4.2'
     implementation 'org.jsoup:jsoup:1.8.3'
-    //Lombok
-    implementation 'org.projectlombok:lombok:1.16.20'
-    annotationProcessor 'org.projectlombok:lombok:1.16.20'
 }
 apply plugin: 'com.google.gms.google-services'
