@@ -1,49 +1,38 @@
 package ru.alexsuvorov.paistewiki.model;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
 
-import lombok.Getter;
-import lombok.Setter;
+import java.util.Objects;
 
 @Entity(tableName = "cymbalseries")
 public class CymbalSeries {
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "cymbalseries_id")
-    @Getter @Setter
     public int cymbalseries_id;   //Ид серии
+    @NonNull
     @ColumnInfo(name = "cymbalseries_name")
-    @Getter @Setter @NonNull
     public String cymbalName;      //Название серии
     @ColumnInfo(name = "cymbalseries_subname")
-    @Getter @Setter
     public String cymbalSubName;   //Девиз серии
     @ColumnInfo(name = "cymbalseries_singleimageuri")
-    @Getter @Setter
     public String cymbalImage;   //Картинка одной тарелки
     @ColumnInfo(name = "cymbalseries_imageuri")
-    @Getter @Setter
     public String seriesImage;   //Картинка серии
     @ColumnInfo(name = "cymbalseries_description")
-    @Getter @Setter
     public String seriesDescription;   //Описание серии
     @ColumnInfo(name = "cymbalseries_isproduced")
-    @Getter @Setter
     public int seriesIsProduced = 1;   //В производстве?
     @ColumnInfo(name = "cymbalseries_description_application")
-    @Getter @Setter
     public String seriesDescriptionApplication;
     @ColumnInfo(name = "cymbalseries_description_since")
-    @Getter @Setter
     public String seriesDescriptionSince;
     @ColumnInfo(name = "cymbalseries_description_sound")
-    @Getter @Setter
     public String seriesDescriptionSound;
     @ColumnInfo(name = "cymbalseries_description_alloy")
-    @Getter @Setter
     public String seriesDescriptionAlloy;
 
     public CymbalSeries() {
@@ -55,7 +44,7 @@ public class CymbalSeries {
         if (o == null || getClass() != o.getClass()) return false;
         CymbalSeries cymbalSeries = (CymbalSeries) o;
         if (cymbalseries_id != cymbalSeries.cymbalseries_id) return false;
-        return cymbalName != null ? cymbalName.equals(cymbalSeries.cymbalName) : cymbalSeries.cymbalName == null;
+        return Objects.equals(cymbalName, cymbalSeries.cymbalName);
     }
 
     @Override
@@ -65,6 +54,7 @@ public class CymbalSeries {
         return result;
     }
 
+    @NonNull
     @Override
     public String toString() {
         return "CymbalSeries{" +
@@ -76,5 +66,93 @@ public class CymbalSeries {
                 ", seriesDescription='" + seriesDescription + '\'' +
                 ", seriesIsProduced=" + seriesIsProduced +
                 '}';
+    }
+
+    public int getCymbalseries_id() {
+        return cymbalseries_id;
+    }
+
+    public void setCymbalseries_id(int cymbalseries_id) {
+        this.cymbalseries_id = cymbalseries_id;
+    }
+
+    public String getCymbalName() {
+        return cymbalName;
+    }
+
+    public void setCymbalName(String cymbalName) {
+        this.cymbalName = cymbalName;
+    }
+
+    public String getCymbalSubName() {
+        return cymbalSubName;
+    }
+
+    public void setCymbalSubName(String cymbalSubName) {
+        this.cymbalSubName = cymbalSubName;
+    }
+
+    public String getCymbalImage() {
+        return cymbalImage;
+    }
+
+    public void setCymbalImage(String cymbalImage) {
+        this.cymbalImage = cymbalImage;
+    }
+
+    public String getSeriesImage() {
+        return seriesImage;
+    }
+
+    public void setSeriesImage(String seriesImage) {
+        this.seriesImage = seriesImage;
+    }
+
+    public String getSeriesDescription() {
+        return seriesDescription;
+    }
+
+    public void setSeriesDescription(String seriesDescription) {
+        this.seriesDescription = seriesDescription;
+    }
+
+    public int getSeriesIsProduced() {
+        return seriesIsProduced;
+    }
+
+    public void setSeriesIsProduced(int seriesIsProduced) {
+        this.seriesIsProduced = seriesIsProduced;
+    }
+
+    public String getSeriesDescriptionApplication() {
+        return seriesDescriptionApplication;
+    }
+
+    public void setSeriesDescriptionApplication(String seriesDescriptionApplication) {
+        this.seriesDescriptionApplication = seriesDescriptionApplication;
+    }
+
+    public String getSeriesDescriptionSince() {
+        return seriesDescriptionSince;
+    }
+
+    public void setSeriesDescriptionSince(String seriesDescriptionSince) {
+        this.seriesDescriptionSince = seriesDescriptionSince;
+    }
+
+    public String getSeriesDescriptionSound() {
+        return seriesDescriptionSound;
+    }
+
+    public void setSeriesDescriptionSound(String seriesDescriptionSound) {
+        this.seriesDescriptionSound = seriesDescriptionSound;
+    }
+
+    public String getSeriesDescriptionAlloy() {
+        return seriesDescriptionAlloy;
+    }
+
+    public void setSeriesDescriptionAlloy(String seriesDescriptionAlloy) {
+        this.seriesDescriptionAlloy = seriesDescriptionAlloy;
     }
 }
