@@ -1,28 +1,20 @@
 package ru.alexsuvorov.paistewiki.model;
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.Ignore;
-import android.arch.persistence.room.PrimaryKey;
-import android.support.annotation.NonNull;
-
-import lombok.Getter;
-import lombok.Setter;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.Ignore;
+import androidx.room.PrimaryKey;
 
 @Entity(tableName = "support_table")
 public class SupportModel {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "support_id")
-    @Getter @Setter @NonNull
     public long support_id = 1;
     @ColumnInfo(name = "support_title")
-    @Getter @Setter
     public String title;
     @ColumnInfo(name = "support_text")
-    @Getter @Setter
     public String text;
     @ColumnInfo(name = "support_image")
-    @Getter @Setter
     public String supportImage;
 
     public SupportModel() {
@@ -33,6 +25,38 @@ public class SupportModel {
         this.support_id = support_id;
         this.title = title;
         this.text = text;
+        this.supportImage = supportImage;
+    }
+
+    public long getSupport_id() {
+        return support_id;
+    }
+
+    public void setSupport_id(long support_id) {
+        this.support_id = support_id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getText() {
+        return text;
+    }
+
+    public void setText(String text) {
+        this.text = text;
+    }
+
+    public String getSupportImage() {
+        return supportImage;
+    }
+
+    public void setSupportImage(String supportImage) {
         this.supportImage = supportImage;
     }
 }
