@@ -2,31 +2,28 @@ package ru.alexsuvorov.paistewiki.activity;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
 import android.view.MenuItem;
 
-import com.google.android.gms.ads.MobileAds;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+
+import com.google.android.material.navigation.NavigationView;
 
 import ru.alexsuvorov.paistewiki.App;
 import ru.alexsuvorov.paistewiki.R;
 import ru.alexsuvorov.paistewiki.fragments.AboutAppFragment;
 import ru.alexsuvorov.paistewiki.fragments.CymbalsFragment;
 import ru.alexsuvorov.paistewiki.fragments.InstagramViewerFragment;
-import ru.alexsuvorov.paistewiki.fragments.NewsFragment;
 import ru.alexsuvorov.paistewiki.fragments.SupportFragment;
 import ru.alexsuvorov.paistewiki.tools.AppPreferences;
 
-public class ContentActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener {
+public class ContentActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
 
     private Fragment fragment;
     private FragmentManager fragmentManager;
@@ -42,11 +39,10 @@ public class ContentActivity extends AppCompatActivity
         fragmentManager = getSupportFragmentManager();
         ((App) getApplication()).setLocale();
 
-        MobileAds.initialize(getApplicationContext(), "ca-app-pub-3940256099942544/6300978111");
         setContentView(R.layout.activity_main);
 
         if (savedInstanceState == null) {
-            fragment = new NewsFragment();
+            fragment = new CymbalsFragment();
             FragmentManager fragmentManager = getSupportFragmentManager();
             fragmentManager.beginTransaction().replace(R.id.container, fragment).commit();
         } else {
@@ -62,17 +58,6 @@ public class ContentActivity extends AppCompatActivity
         toggle.syncState();
 
         NavigationView navigationView = findViewById(R.id.nav_view);
-        /*
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
-    if (navigationView != null) {
-        Menu menu = navigationView.getMenu();
-        menu.findItem(R.id.nav_profile).setTitle("My Account");
-        menu.findItem(R.id.nav_mng_task).setTitle("Control Task");
-        //menu.findItem(R.id.nav_pkg_manage).setVisible(false);//In case you want to remove menu item
-        navigationView.setNavigationItemSelectedListener(this);
-    }
-         */
-
         navigationView.setNavigationItemSelectedListener(this);
         navigationView.getMenu().clear();
         navigationView.inflateMenu(R.menu.menu_activity_main);
@@ -89,20 +74,13 @@ public class ContentActivity extends AppCompatActivity
         }
     }
 
-    /*@Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        if (!drawer.isDrawerOpen(GravityCompat.START)) {
-            return true;
-        }
-        return super.onCreateOptionsMenu(menu);
-    }*/
-
     public boolean onNavigationItemSelected(@NonNull MenuItem item) {
         int id = item.getItemId();
-        if (id == R.id.nav_news) {
+        /*if (id == R.id.nav_news) {
             fragment = new NewsFragment();
             fragmentManager.beginTransaction().replace(R.id.container, fragment).commit();
-        } else if (id == R.id.nav_cymbals) {
+        } else */
+        if (id == R.id.nav_cymbals) {
             fragment = new CymbalsFragment();
             fragmentManager.beginTransaction().replace(R.id.container, fragment).commit();
         } else if (id == R.id.nav_support) {
@@ -111,7 +89,7 @@ public class ContentActivity extends AppCompatActivity
         } else if (id == R.id.nav_instagram) {
             fragment = new InstagramViewerFragment();
             fragmentManager.beginTransaction().replace(R.id.container, fragment).commit();
-        }else if (id == R.id.nav_about) {
+        } else if (id == R.id.nav_about) {
             fragment = new AboutAppFragment();
             fragmentManager.beginTransaction().replace(R.id.container, fragment).commit();
         }
