@@ -1,12 +1,12 @@
 package ru.alexsuvorov.paistewiki.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import ru.alexsuvorov.paistewiki.App;
 import ru.alexsuvorov.paistewiki.R;
@@ -18,7 +18,7 @@ public class SeriesDescriptionActivity extends AppCompatActivity {
     Toolbar toolbar;
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         ((App) getApplication()).setLocale();
         setContentView(R.layout.fragment_cymbal_description);
