@@ -1,9 +1,10 @@
 package ru.alexsuvorov.paistewiki.activity.support;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import ru.alexsuvorov.paistewiki.App;
 import ru.alexsuvorov.paistewiki.R;
