@@ -1,10 +1,7 @@
 package ru.alexsuvorov.paistewiki.adapter;
 
-import android.app.FragmentManager;
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,13 +9,15 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import java.util.List;
 
 import ru.alexsuvorov.paistewiki.R;
 import ru.alexsuvorov.paistewiki.activity.support.SupportAnatomyActivity;
 import ru.alexsuvorov.paistewiki.db.AppDatabase;
 import ru.alexsuvorov.paistewiki.db.dao.SupportDao;
-import ru.alexsuvorov.paistewiki.fragments.InstagramViewerFragment;
 import ru.alexsuvorov.paistewiki.model.SupportModel;
 import ru.alexsuvorov.paistewiki.tools.Utils;
 
@@ -28,9 +27,9 @@ public class SupportAdapter extends RecyclerView.Adapter<SupportAdapter.ViewHold
         void onClick(int position);
     }
 
-    private List<SupportModel> supportModelList;
-    private SupportCallback listener;
-    private Context context;
+    private final List<SupportModel> supportModelList;
+    private final SupportCallback listener;
+    private final Context context;
 
     public SupportAdapter(List<SupportModel> supportModelList, Context context, SupportCallback listener) {
         this.supportModelList = supportModelList;
@@ -87,10 +86,10 @@ public class SupportAdapter extends RecyclerView.Adapter<SupportAdapter.ViewHold
     }
 
     class ViewHolder extends RecyclerView.ViewHolder {
-        private ImageView supportImage;
-        private TextView supportTitle;
-        private TextView supportText;
-        private LinearLayout supportLayout;
+        private final ImageView supportImage;
+        private final TextView supportTitle;
+        private final TextView supportText;
+        private final LinearLayout supportLayout;
 
         ViewHolder(View itemView) {
             super(itemView);
